+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next version
 + Add CircleCI
++ Add balance api and reducer
 
 ## 0.1.0
 + First commit